+++ src/main/java/video/api/java/sdk/infrastructure/pagination/PageDeserializer.java
@@ -1,7 +1,7 @@
 package video.api.java.sdk.infrastructure.pagination;
 
-import org.json.JSONException;
-import org.json.JSONObject;
+import kong.unirest.json.JSONException;
+import kong.unirest.json.JSONObject;
 import video.api.java.sdk.domain.pagination.Page;
 import video.api.java.sdk.infrastructure.unirest.serializer.JsonDeserializer;