+++ src/test/java/video/api/java/sdk/infrastructure/unirest/EmptySerializer.java
@@ -1,7 +1,7 @@
 package video.api.java.sdk.infrastructure.unirest;
 
-import org.json.JSONException;
-import org.json.JSONObject;
+import kong.unirest.json.JSONException;
+import kong.unirest.json.JSONObject;
 import video.api.java.sdk.infrastructure.unirest.serializer.JsonSerializer;
 
 public class EmptySerializer<T> implements JsonSerializer<T> {